+++ examples/query_history.rs
@@ -4,7 +4,7 @@ use anyhow::Result;
 use futures::StreamExt;
 use scylla::history::{HistoryCollector, StructuredHistory};
 use scylla::query::Query;
-use scylla::transport::session::Session;
+use scylla::session::Session;
 use scylla::SessionBuilder;
 use std::env;
 use std::sync::Arc;
