+++ examples/tls.rs
@@ -1,5 +1,5 @@
 use anyhow::Result;
-use scylla::transport::session::{IntoTypedRows, Session};
+use scylla::session::{IntoTypedRows, Session};
 use scylla::SessionBuilder;
 use std::env;
 use std::fs;
