+++ scylla/src/transport/connection.rs
@@ -57,10 +57,10 @@ use crate::frame::{
     FrameParams, SerializedRequest,
 };
 use crate::query::Query;
+use crate::session::IntoTypedRows;
 use crate::sharding::ShardInfo;
 use crate::statement::prepared_statement::PreparedStatement;
 use crate::statement::Consistency;
-use crate::transport::session::IntoTypedRows;
 use crate::transport::Compression;
 use crate::QueryResult;
