+++ scylla/src/transport/cql_types_test.rs
@@ -5,9 +5,9 @@ use crate::frame::value::Counter;
 use crate::frame::value::Value;
 use crate::frame::value::{Date, Time, Timestamp};
 use crate::macros::{FromUserType, IntoUserType};
+use crate::session::IntoTypedRows;
+use crate::session::Session;
 use crate::test_utils::create_new_session_builder;
-use crate::transport::session::IntoTypedRows;
-use crate::transport::session::Session;
 use crate::utils::test_utils::unique_keyspace_name;
 use bigdecimal::BigDecimal;
 use chrono::{Duration, NaiveDate};
