+++ scylla/tests/integration/consistency.rs
@@ -4,9 +4,9 @@ use scylla::execution_profile::{ExecutionProfileBuilder, ExecutionProfileHandle}
 use scylla::load_balancing::{DefaultPolicy, LoadBalancingPolicy, RoutingInfo};
 use scylla::prepared_statement::PreparedStatement;
 use scylla::retry_policy::FallthroughRetryPolicy;
+use scylla::session::Session;
 use scylla::sharding::Token;
 use scylla::test_utils::unique_keyspace_name;
-use scylla::transport::session::Session;
 use tokio::sync::mpsc::{self, UnboundedReceiver, UnboundedSender};
 
 use scylla::statement::batch::BatchStatement;
