+++ scylla/tests/integration/retries.rs
@@ -1,7 +1,7 @@
 use crate::utils::test_with_3_node_cluster;
 use scylla::retry_policy::FallthroughRetryPolicy;
+use scylla::session::Session;
 use scylla::speculative_execution::SimpleSpeculativeExecutionPolicy;
-use scylla::transport::session::Session;
 use scylla::ExecutionProfile;
 use scylla::SessionBuilder;
 use scylla::{query::Query, test_utils::unique_keyspace_name};
